requirements.txt
requests
python-dotenv
langchain
langchain-groq
google-cloud-storage
fastapi
pydantic[email]
uvicorn
openai